requirements.txt
# Core dependencies (optimized for minimal footprint)
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0